# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Getting information",
        "version": "1.0.0"
    },
    "paths": {
        "/": {},
        "/customers/{customer}": {
            "get": {
                "tags": [
                    "Show a Customer"
                ],
                "summary": "Get one of users",
                "description": "To show a user informaion in this form by using route model binding",
                "operationId": "customer.show",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "an instance of Customer model(I have used route model binding)",
                        "required": true,
                        "schema": {
                            "type": "Customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Removing"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "an instance of Customer model(I have used route model binding)",
                        "required": true,
                        "schema": {
                            "type": "Customer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/customers": {
            "post": {
                "tags": [
                    "Registering"
                ],
                "summary": "Register a new user",
                "description": "To register user informaion in this form and store it in the database",
                "operationId": "customer.store",
                "parameters": [
                    {
                        "name": "Firstname",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Lastname",
                        "in": "query",
                        "description": "User's last name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "DateOfBirth",
                        "in": "query",
                        "description": "User's birthday",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "BankAccountNumber",
                        "in": "query",
                        "description": "User's bank account number",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/customers/{customers}": {
            "put": {
                "tags": [
                    "Modifying"
                ],
                "summary": "Modifying Info",
                "description": "To modify user informaion in this form and store it in the database",
                "operationId": "customer.update",
                "parameters": [
                    {
                        "name": "Firstname",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Lastname",
                        "in": "query",
                        "description": "User's last name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "DateOfBirth",
                        "in": "query",
                        "description": "User's birthday",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "BankAccountNumber",
                        "in": "query",
                        "description": "User's bank account number",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        }
    }
}